+++ kipper/core/src/compiler/ast/nodes/declarations/type-declaration/interface-declaration/interface-declaration-semantics.ts
@@ -5,6 +5,7 @@
 import type { TypeDeclarationSemantics } from "../type-declaration-semantics";
 import type { InterfaceMemberDeclaration } from "./interface-member-declaration";
 import { InterfaceMemberDeclarationSemantics } from "./interface-member-declaration";
+import { ProcessedType } from "../../../../../semantics";
 
 /**
  * Semantics for AST Node {@link ClassDeclaration}.